WIP: refactor ingredientsList, extract section and item components
